requirements.txt
# Core dependencies
python-dotenv>=0.19.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2

# Security
cryptography>=3.4.7
PyJWT>=2.1.0
python-jose>=3.3.0
passlib>=1.7.4

# Data handling
pandas>=1.3.0
numpy>=1.21.0
requests>=2.26.0
aiohttp>=3.8.0

# NLP and ML
nltk>=3.6.3
scikit-learn>=0.24.2
transformers>=4.9.2
torch>=1.9.0

# Database
sqlalchemy>=1.4.23
alembic>=1.7.1
psycopg2-binary>=2.9.1

# Monitoring and logging
prometheus-client>=0.11.0
python-json-logger>=2.0.2

# Testing
pytest>=6.2.5
pytest-asyncio>=0.15.1
pytest-cov>=2.12.1

# Development
black>=21.7b0
flake8>=3.9.2
mypy>=0.910
isort>=5.9.3 